requirements.txt
astroid==1.6.5
certifi==2018.4.16
chardet==3.0.4
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.2
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
urllib3==1.23
wrapt==1.10.11
